pyproject.toml
[tool.poetry]
name = "pageplus"
version = "0.1.0"
description = ""
authors = ["JKamlah <jan.kamlah@uni-mannheim.de>"]
readme = "README.md"

[tool.poetry.scripts]
pageplus = "pageplus.main:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
lxml = "^5.1.0"
shapely = "^2.0.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
urllib3 = "<2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
